requirements.txt
flask
python-dotenv
pypdf
google-genai
numpy
faiss-cpu
langchain
langchain-google-genai
langchain-community
